gave the View default implementations for some methods so that the CLIView and GUIView are less crowded with uneccessary @Overrides
